Fjernet depricated kode - hentSimulering og lagreTilbakekreving på vedtak
